requirements.txt
Flask
mysql-connector-python
ping3